backend/requirements.txt
argon2-cffi
flask
flask-cors
pydantic
pymysql[rsa]
